requirements.txt
edge-tts
google-generativeai
moviepy
numpy
Pillow
python-dotenv
requests
torch
whisperx